requirements.txt
Flask
Flask-cors
opencv-python
ffmpeg-python
ultralytics